Fix: UserRespository에 queryDsl 적용 (maxUserIdx 메소드만 - 나머지도 적용 예정)

+++ src/main/java/com/jm/portfolio/domain/users/service/impl/EmailSignUpServiceImpl.java
@@ -36,7 +36,7 @@ public class EmailSignUpServiceImpl implements SignUpService {
         user.hashPassword(passwordEncoder);
         userRepository.save(user);
 
-        int maxIdx = userRepository.maxUserIdx();
+        long maxIdx = userRepository.maxUserIdx();
         Role role = new Role(newUser.getCreatedIp(), newUser.getLastUpdatedIp(), maxIdx, "USER");
         roleRepository.save(role);
